requirements.txt
asgiref
Django
djangorestframework
pillow
psycopg2-binary
python-dotenv
pytz
sqlparse
tzdata
django-filter